Moves outputNode to separate file : outputNode.go
